refactor(armies-lists): Fix tests after relations updates
