+++ src/army/army.module.ts
@@ -15,7 +15,6 @@ import { EquipmentUnitTroop } from "./unit/troop/equipment/equipment-unit-troop.
 import { SpecialRuleUnitTroop } from "./unit/troop/special-rule/special-rule-unit-troop.entity";
 import { ArmyService } from "./army.service";
 import { ArmyController } from "./army.controller";
-// import { UnitService } from "@army/unit/troop/unit.service";
 import { UnitModule } from "@army/unit/unit.module";
 
 @Module({
