+++ src/database/migration/1690531446154-ArmyListUnitRelationsRefactoring.ts
@@ -1,4 +1,4 @@
-import { MigrationInterface, QueryRunner, Table, TableColumn, TableForeignKey } from "typeorm";
+import { MigrationInterface, QueryRunner, Table, TableColumn } from "typeorm";
 
 const ARMY_LIST_UNITS_TROOP_IDS_COLUMN: TableColumn = new TableColumn(
     {
@@ -22,18 +22,6 @@ const ARMY_LIST_UNITS_TROOPS_TABLE: Table = new Table({
     ]
 });
 
-const ARMY_LIST_UNITS_TROOPS_TABLE_FOREIGN_KEYS: TableForeignKey[] = [
-    new TableForeignKey({
-        columnNames: ["army_list_unit_id"],
-        referencedColumnNames: ["id"],
-        referencedTableName: "army_list_units"
-    }), new TableForeignKey({
-        columnNames: ["troop_id"],
-        referencedColumnNames: ["id"],
-        referencedTableName: "troops"
-    })
-];
-
 export class ArmyListUnitRelationsRefactoring1690531446154 implements MigrationInterface {
 
     public async up(queryRunner: QueryRunner): Promise<void> {
@@ -41,7 +29,6 @@ export class ArmyListUnitRelationsRefactoring1690531446154 implements MigrationI
         await queryRunner.dropColumn("army_list_units", ARMY_LIST_UNITS_TROOP_IDS_COLUMN);
         // create army list units troops table
         await queryRunner.createTable(ARMY_LIST_UNITS_TROOPS_TABLE);
-        // await queryRunner.createForeignKeys(ARMY_LIST_UNITS_TROOPS_TABLE, ARMY_LIST_UNITS_TROOPS_TABLE_FOREIGN_KEYS);
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
