+++ tests/api/army-list/lookup.spec.ts
@@ -1,0 +1,73 @@
+import { Test, TestingModule } from "@nestjs/testing";
+import { AppModule } from "../../../src/app.module";
+import { TestsHelper } from "../../tests.helper";
+import * as request from "supertest";
+import { faker } from "@faker-js/faker";
+import { HttpStatus, INestApplication } from "@nestjs/common";
+import { ARMY1 } from "../../fixtures/army-list/armies-lists";
+
+jest.setTimeout(25000);
+
+const USERNAME = faker.internet.userName();
+const EMAIL = faker.internet.email();
+const PASSWORD = faker.internet.password();
+
+const USERNAME1 = faker.internet.userName();
+const EMAIL1 = faker.internet.email();
+const PASSWORD1 = faker.internet.password();
+
+let app: INestApplication;
+let token: string;
+let token1: string;
+
+describe("armies-lists/create", () => {
+    beforeAll(async () => {
+        const module: TestingModule = await Test.createTestingModule({
+            imports: [AppModule]
+        }).compile();
+
+        app = module.createNestApplication();
+        await app.init();
+
+        token = await TestsHelper.createAccountAndGetToken(app.getHttpServer(), USERNAME, EMAIL,
+            PASSWORD);
+        token1 = await TestsHelper.createAccountAndGetToken(app.getHttpServer(), USERNAME1, EMAIL1,
+            PASSWORD);
+
+        await request(app.getHttpServer())
+            .post(TestsHelper.ARMIES_LISTS_CREATE_ROUTE)
+            .set("Authorization", `Bearer ${token}`).send(ARMY1);
+    });
+
+    afterAll(async () => {
+        const res = await request(app.getHttpServer())
+            .get(TestsHelper.ARMIES_LISTS_LOOKUP_ROUTE)
+            .set("Authorization", `Bearer ${token}`);
+
+        for (const a of res.body) {
+            await request(app.getHttpServer())
+                .delete(`${TestsHelper.ARMIES_LISTS_DELETE_ROUTE}/${a.id}`)
+                .set("Authorization", `Bearer ${token}`);
+        }
+        await TestsHelper.deleteAccount(app.getHttpServer(), token);
+        await TestsHelper.deleteAccount(app.getHttpServer(), token1);
+    });
+
+    it("lookup: basic lookup - then return armies lists credentials", async () => {
+        const res = await request(app.getHttpServer())
+            .get(TestsHelper.ARMIES_LISTS_LOOKUP_ROUTE)
+            .set("Authorization", `Bearer ${token}`);
+
+        expect(res.status).toEqual(HttpStatus.OK);
+        expect(res.body).toBeDefined();
+    });
+
+    it("lookup: with invalid token - then return 401 (unauthorized)", async () => {
+        const res = await request(app.getHttpServer())
+            .get(TestsHelper.ARMIES_LISTS_LOOKUP_ROUTE)
+            .set("Authorization", `Bearer abcd`);
+
+        expect(res.status).toEqual(HttpStatus.UNAUTHORIZED);
+    });
+
+});
